+++ lib/screens/onboarding/onboarding_screen.dart
@@ -39,7 +39,7 @@ class _OnboardingScreenState extends State<OnboardingScreen> {
       appBar: AppBar(
         backgroundColor: Colors.transparent,
         elevation: 0.0,
-        brightness: Brightness.light,
+        systemOverlayStyle: SystemUiOverlayStyle.dark,
       ),
       body: Stack(alignment: AlignmentDirectional.bottomCenter, children: [
         PageView.builder(