+++ lib/screens/projects/project_item.dart
@@ -11,9 +11,10 @@ class ProjectOrganizationItem extends ProjectItem {
 }
 
 class ProjectProjectItem extends ProjectItem {
-  ProjectProjectItem(
-      this.organizationSlug, this.projectSlug, this.isBookmarked);
+  ProjectProjectItem(this.platform, this.organizationSlug, this.projectSlug,
+      this.isBookmarked);
 
+  final String? platform;
   final String organizationSlug;
   final String projectSlug;
   final bool isBookmarked;