+++ lib/screens/projects/projects_screen.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
+import 'package:sentry_mobile/utils/platform_icons.dart';
 
 import '../../redux/actions.dart';
 import '../../redux/state/app_state.dart';
@@ -29,8 +30,7 @@ class _ProjectsScreenState extends State<ProjectsScreen> {
 
   Widget _content(ProjectsScreenViewModel viewModel) {
     return Scaffold(
-      appBar: AppBar(
-          title: Text('Bookmarked Projects'), brightness: Brightness.dark),
+      appBar: AppBar(title: Text('Bookmarked Projects')),
       body: viewModel.items.isEmpty
           ? Center(
               child: Padding(
@@ -69,6 +69,7 @@ class _ProjectsScreenState extends State<ProjectsScreen> {
                           .bodyText1
                           ?.apply(color: SentryColors.revolver),
                     ),
+                    leading: _platformImage(item.platform),
                     trailing: Icon(
                       item.isBookmarked ? Icons.star : Icons.star_border,
                       color: item.isBookmarked
@@ -83,6 +84,32 @@ class _ProjectsScreenState extends State<ProjectsScreen> {
                   throw Exception('Unknown list item type.');
                 }
               }),
+    );
+  }
+
+  Widget? _platformImage(String? platform) {
+    const width = 22.0;
+    const height = 22.0;
+    Widget? platformImage;
+    if (platform != null) {
+      platformImage = PlatformIcons.svgPicture(platform, width, height);
+    } else {
+      platformImage = Container(
+        width: width,
+        height: height,
+      );
+    }
+    return Padding(
+      padding: const EdgeInsets.only(left: 8),
+      child: ClipRRect(
+        borderRadius: BorderRadius.all(Radius.circular(2)),
+        child: Container(
+          width: width,
+          height: height,
+          color: SentryColors.snuff,
+          child: platformImage,
+        ),
+      ),
     );
   }
 }