+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Unreleased
 
-* Bump Sentry Flutter 6.0.0 (#211)
 * Update Metadata (#216)
+* Show platform icons in project picker (#214)
+* Bump Sentry Flutter 6.0.0 (#211)
 
 # 1.0.3